+++ oxytcmri/tests/unit/infrastructure/test_importers.py
@@ -1,7 +1,7 @@
 import pytest
 
-from oxytcmri.interface.importers import CSVAtlasImporter, NiftiFoldersImporter
-from oxytcmri.infrastructure.importers.csv import CSVCenterImporter
+from oxytcmri.infrastructure.importers.nifti_folders import NiftiFoldersImporter
+from oxytcmri.infrastructure.importers.csv import CSVCenterImporter, CSVAtlasImporter
 from oxytcmri.tests.fixtures import path_to_test_data_folder
 from oxytcmri.tests.unit.domain.mocks import MockEmptyCenterRepository, MockAtlasRepository, \
     MockInMemoryEmptyMRIRepository, \
